// ams-fe/src/app/model/appUserDto.ts
/**
 * Demo REST API
 * User Management API.
 *
 * OpenAPI spec version: API v1.0.0
 * Contact: nhtoi@tma.com.vn
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface AppUserDto { 
    address?: string;
    avatar?: string;
    birthDate?: string;
    email?: string;
    firstName?: string;
    fullName?: string;
    gender?: string;
    id?: number;
    lastName?: string;
    password?: string;
    phone?: string;
    role?: string;
    status?: string;
    username?: string;
}